Ktor
Ktor is an asynchronous open source framework for creating microservices and web applications. It was developed with Kotlin by Jetbrains. It’s easy to install and use, It’s extensible if you want to add a step to the execution pipeline. Its asynchronous property allows it to receive multiple requests thanks to coroutines and it is also multiplatform.